smart-resume-builder/requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-allauth==0.52.0
Pillow==10.0.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pdfkit==1.0.0
python-docx==0.8.11
openai==0.27.0
nltk==3.7
django-cors-headers==3.14.0
django-filter==23.2.0
django-extensions==3.2.1
pytest==7.3.1
pytest-django==4.5.2